изменения: - фикс ошибки открытия панели магазина (#105998060) - в интерфейс IInit добавил булевую пропертю Initialized; - GameClient теперь биндится через DI; - TestUtilsHelper для теста любых методов в коде; - фикс бага, когда персонаж может случайно совершить движение после поворота; - отправляем аналитики в старый и новый дашборд юнити одновременно; - аутентификация в игровые сервисы корректно отрабатывает как на iOS, так и на Android; - доска лидеров корректно отрабатывает как на iOS, так и на Android; - установка и получение значений доски лидеров корректно отрабатывает на iOS - менеджер магазина для iOS: AppleSAShopManager; - корректная проверка наличия интернет соединения в NetworkUtils;

+++ Client/Assets/Scripts/Entities/ScoresEntity.cs
@@ -1,15 +1,14 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using DI.Extensions;
-using Managers;
 
 namespace Entities
 {
-    public class ScoresEntity : EntityBase<Dictionary<int, long>>
+    public class ScoresEntity : EntityBase<Dictionary<ushort, long>>
     {
         public ScoresEntity()
         {
-            Value = new Dictionary<int, long>();
+            Value = new Dictionary<ushort, long>();
         }
 
         public long? GetFirstScore()
@@ -19,7 +18,7 @@ namespace Entities
             return null;
         }
         
-        public long? GetScore(int _Key)
+        public long? GetScore(ushort _Key)
         {
             return Value.GetSafe(_Key, out _);
         }
